app/requirements.txt
flask
Flask-Session
redis
setuptools
psycopg2-binary
setuptools
sqlalchemy
pandas
uuid
python-dotenv
flask_migrate
flask_sqlalchemy
werkzeug
isodate
google-api-python-client
debugpy
gunicorn
PyYAML
Flask-Session